opencv-flask/requirements.txt
camera==1.3.0
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.16.2
opencv-python==4.0.1.24
Pillow==6.0.0
Werkzeug==1.0.1
